refactor: move agente-edital sections to shared components
